+++ src/utils/GdiPlusUtil.h
@@ -10,15 +10,15 @@
 // note: must write "using namespace Gdiplus;" before #include "GdiPlusUtil.h"
 // this is to make sure we don't accidentally do that just by including this file
 
-typedef RectF (* TextMeasureAlgorithm)(Graphics *g, Font *f, const WCHAR *s, size_t len);
+typedef RectF (* TextMeasureAlgorithm)(Graphics *g, Font *f, const WCHAR *s, int len);
 
-RectF    MeasureTextAccurate(Graphics *g, Font *f, const WCHAR *s, size_t len);
-RectF    MeasureTextAccurate2(Graphics *g, Font *f, const WCHAR *s, size_t len);
-RectF    MeasureTextStandard(Graphics *g, Font *f, const WCHAR *s, size_t len);
-RectF    MeasureTextQuick(Graphics *g, Font *f, const WCHAR *s, size_t len);
+RectF    MeasureTextAccurate(Graphics *g, Font *f, const WCHAR *s, int len);
+RectF    MeasureTextAccurate2(Graphics *g, Font *f, const WCHAR *s, int len);
+RectF    MeasureTextStandard(Graphics *g, Font *f, const WCHAR *s, int len);
+RectF    MeasureTextQuick(Graphics *g, Font *f, const WCHAR *s, int len);
 RectF    MeasureText(Graphics *g, Font *f, const WCHAR *s, size_t len=-1, TextMeasureAlgorithm algo=NULL);
 REAL     GetSpaceDx(Graphics *g, Font *f, TextMeasureAlgorithm algo=NULL);
-int      StringLenForWidth(Graphics *g, Font *f, const WCHAR *s, size_t len, float dx, TextMeasureAlgorithm algo=NULL);
+size_t   StringLenForWidth(Graphics *g, Font *f, const WCHAR *s, size_t len, float dx, TextMeasureAlgorithm algo=NULL);
 void     DrawCloseButton(DRAWITEMSTRUCT *dis);
 
 void     GetBaseTransform(Matrix& m, RectF pageRect, float zoom, int rotation);
